pyproject.toml
[tool.poetry]
name = "synthetics_cdk"
version = "0.1.0"
description = "Generate Datadog synthetics tests using CDK"
authors = ["Datadog <support@datadoghq.com>"]
license = "Apache-2.0"

[tool.poetry.dependencies]
python = "~3.10"
click = "~8.1.3"
cdktf = ">=0.11"
cdktf-cdktf-provider-datadog = ">=2.0.0"
PyYaml = "~6.0"
